+++ blocklets/ai-studio/api/src/libs/security.ts
@@ -1,4 +1,4 @@
-import { auth } from '@blocklet/sdk/lib/middlewares';
+import middlewares from '@blocklet/sdk/lib/middlewares';
 import { getVerifyData, verify } from '@blocklet/sdk/lib/util/verify-sign';
 import { NextFunction, Request, Response } from 'express';
 
@@ -7,15 +7,15 @@ import logger from './logger';
 
 export const ADMIN_ROLES = ['owner', 'admin'];
 
-export const ensureAdmin = auth({ roles: ADMIN_ROLES });
+export const ensureAdmin = middlewares.auth({ roles: ADMIN_ROLES });
 
 // dynamic permission check
 export const ensurePromptsAdmin = (req: Request, res: Response, next: NextFunction) =>
-  auth({ roles: Config.serviceModePermissionMap.ensurePromptsAdminRoles })(req, res, next);
+  middlewares.auth({ roles: Config.serviceModePermissionMap.ensurePromptsAdminRoles })(req, res, next);
 
 // dynamic permission check
 export const ensurePromptsEditor = (req: Request, res: Response, next: NextFunction) =>
-  auth({ roles: Config.serviceModePermissionMap.ensurePromptsEditorRoles })(req, res, next);
+  middlewares.auth({ roles: Config.serviceModePermissionMap.ensurePromptsEditorRoles })(req, res, next);
 
 export const isRefReadOnly = ({
   ref,
@@ -90,5 +90,5 @@ export function ensureComponentCallOrRolesMatch(req: Request, roles?: string[]) 
 }
 
 export function ensureComponentCallOrAuth() {
-  return ensureComponentCallOr(auth());
+  return ensureComponentCallOr(middlewares.auth());
 }