+++ blocklets/ai-studio/api/src/routes/project.ts
@@ -26,7 +26,7 @@ import { AIGNE_RUNTIME_COMPONENT_DID, NFT_BLENDER_COMPONENT_DID } from '@blockle
 import { Map, getYjsValue } from '@blocklet/co-git/yjs';
 import { call } from '@blocklet/sdk/lib/component';
 import config from '@blocklet/sdk/lib/config';
-import { user } from '@blocklet/sdk/lib/middlewares';
+import { session } from '@blocklet/sdk/lib/middlewares';
 import { Request, Router } from 'express';
 import { exists } from 'fs-extra';
 import * as git from 'isomorphic-git';
@@ -266,7 +266,7 @@ export interface CreateOrUpdateAgentInputSecretPayload {
 }
 
 export function projectRoutes(router: Router) {
-  router.get('/projects', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.get('/projects', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const list = await Project.findAll({
       where: { ...getProjectWhereConditions(req) },
       order: [
@@ -320,7 +320,7 @@ export function projectRoutes(router: Router) {
     projectId: Joi.string().empty([null, '']),
   });
 
-  router.get('/projects/check-name', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.get('/projects/check-name', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { name, projectId } = await checkProjectNameSchema.validateAsync(req.query, { stripUnknown: true });
     const projects = await Project.findAll({ where: { name, createdBy: req.user.did } });
 
@@ -333,7 +333,7 @@ export function projectRoutes(router: Router) {
     res.json({ ok: filtered.length === 0, project: filtered?.[0] });
   });
 
-  router.get('/template-projects', user(), ensureComponentCallOrPromptsEditor(), async (_req, res) => {
+  router.get('/template-projects', session(), ensureComponentCallOrPromptsEditor(), async (_req, res) => {
     const resourceTemplates = (await resourceManager.getProjects({ type: 'template' })).map((i) => ({
       ...i.project,
       blockletDid: i.blocklet.did,
@@ -342,7 +342,7 @@ export function projectRoutes(router: Router) {
     res.json({ templates: uniqBy([...resourceTemplates], (i) => i.id) });
   });
 
-  router.get('/projects/icons', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.get('/projects/icons', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { did } = req.user!;
 
     const { data } = await call({
@@ -356,7 +356,7 @@ export function projectRoutes(router: Router) {
     res.json({ icons: data?.uploads || [] });
   });
 
-  router.delete('/projects/icon/:id', ensureComponentCallOrPromptsEditor(), user(), async (req, res) => {
+  router.delete('/projects/icon/:id', ensureComponentCallOrPromptsEditor(), session(), async (req, res) => {
     const { did } = req.user!;
     const { data } = await call({
       name: 'image-bin',
@@ -422,7 +422,7 @@ export function projectRoutes(router: Router) {
     working: Joi.boolean().empty(['', null]),
   });
 
-  router.get('/projects/:projectId', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.get('/projects/:projectId', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required param `projectId`');
 
@@ -444,7 +444,7 @@ export function projectRoutes(router: Router) {
 
   router.get(
     '/projects/:projectId/agent-input-secrets',
-    user(),
+    session(),
     ensureComponentCallOrPromptsEditor(),
     async (req, res) => {
       const { projectId } = req.params;
@@ -476,7 +476,7 @@ export function projectRoutes(router: Router) {
 
   router.post(
     '/projects/:projectId/agent-input-secrets',
-    user(),
+    session(),
     ensureComponentCallOrPromptsEditor(),
     async (req, res) => {
       const { did: userId } = req.user!;
@@ -521,7 +521,7 @@ export function projectRoutes(router: Router) {
     }
   );
 
-  router.post('/projects', user(), ensureComponentCallOrPromptsEditor(), checkDeployment, async (req, res) => {
+  router.post('/projects', session(), ensureComponentCallOrPromptsEditor(), checkDeployment, async (req, res) => {
     const {
       blockletDid,
       templateId = projectTemplates[0]?.project?.id,
@@ -590,7 +590,7 @@ export function projectRoutes(router: Router) {
     res.json(project);
   });
 
-  router.post('/projects/import', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.post('/projects/import', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     await checkProjectLimit({ req });
 
     const { name, username, password, description, url } = await importProjectSchema.validateAsync(req.body, {
@@ -699,7 +699,7 @@ export function projectRoutes(router: Router) {
     }
   });
 
-  router.patch('/projects/:projectId', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.patch('/projects/:projectId', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required parameter `projectId`');
 
@@ -761,7 +761,7 @@ export function projectRoutes(router: Router) {
     res.json(project.dataValues);
   });
 
-  router.delete('/projects/:projectId', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.delete('/projects/:projectId', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required params `projectId`');
 
@@ -788,7 +788,7 @@ export function projectRoutes(router: Router) {
     res.json(project);
   });
 
-  router.post('/projects/:projectId/remote', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.post('/projects/:projectId/remote', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required params `projectId`');
 
@@ -816,7 +816,7 @@ export function projectRoutes(router: Router) {
     res.json({});
   });
 
-  router.delete('/projects/:projectId/remote', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.delete('/projects/:projectId/remote', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required params `projectId`');
 
@@ -833,7 +833,7 @@ export function projectRoutes(router: Router) {
     res.json({});
   });
 
-  router.post('/projects/:projectId/remote/push', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.post('/projects/:projectId/remote/push', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { projectId } = req.params;
     if (!projectId) throw new Error('Missing required params `projectId`');
 
@@ -855,7 +855,7 @@ export function projectRoutes(router: Router) {
     res.json({});
   });
 
-  router.post('/projects/:projectId/remote/pull', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.post('/projects/:projectId/remote/pull', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { did: userId, fullName } = req.user!;
 
     const { projectId } = req.params;
@@ -889,7 +889,7 @@ export function projectRoutes(router: Router) {
     res.json({});
   });
 
-  router.post('/projects/:projectId/remote/sync', user(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
+  router.post('/projects/:projectId/remote/sync', session(), ensureComponentCallOrPromptsEditor(), async (req, res) => {
     const { did: userId, fullName } = req.user!;
 
     const { projectId } = req.params;
@@ -1019,7 +1019,7 @@ export function projectRoutes(router: Router) {
 
   router.get(
     '/projects/compare/:projectId/:ref/:agentId',
-    user(),
+    session(),
     ensureComponentCallOrPromptsEditor(),
     async (req, res) => {
       const { projectId, ref, agentId } = req.params;
@@ -1101,7 +1101,7 @@ export function projectRoutes(router: Router) {
 
   router.post(
     '/projects/:projectId/refs/:ref/assets',
-    user(),
+    session(),
     ensureComponentCallOrPromptsEditor(),
     async (req, res) => {
       const { projectId, ref } = req.params;