+++ blocklets/ai-studio/api/src/routes/resource.ts
@@ -16,7 +16,7 @@ import {
 import { copyRecursive } from '@blocklet/ai-runtime/utils/fs';
 import { isNonNullable } from '@blocklet/ai-runtime/utils/is-non-nullable';
 import component, { call } from '@blocklet/sdk/lib/component';
-import { user } from '@blocklet/sdk/lib/middlewares';
+import middlewares from '@blocklet/sdk/lib/middlewares';
 import { Router } from 'express';
 import { pathExists } from 'fs-extra';
 import Joi from 'joi';
@@ -112,7 +112,7 @@ export function resourceRoutes(router: Router) {
     return undefined;
   };
 
-  router.get('/resources/export', user(), ensurePromptsEditor, async (req, res) => {
+  router.get('/resources/export', middlewares.session(), ensurePromptsEditor, async (req, res) => {
     const query = await getExportedResourceQuerySchema.validateAsync(
       {
         ...req.query,