+++ packages/ai-runtime/src/front/components/AgentInputField/index.tsx
@@ -1,11 +1,9 @@
-import Toast from '@arcblock/ux/lib/Toast';
-import AttachFileIcon from '@mui/icons-material/AttachFile';
-import { IconButton, InputAdornment, Stack, TextFieldProps } from '@mui/material';
-import { ComponentType, useRef } from 'react';
+import { TextFieldProps } from '@mui/material';
+import { ComponentType } from 'react';
 
-import { uploadImage } from '../../../api/ai-runtime/image';
 import { Parameter } from '../../../types';
 import BooleanField from './BooleanField';
+import ImageField from './ImageField';
 import LanguageField from './LanguageField';
 import NumberField from './NumberField';
 import RadioField from './RadioField';
@@ -19,10 +17,8 @@ export default function AgentInputField({
 }: {
   readOnly?: boolean;
   parameter: Parameter;
-  onChange: (value: string | number | undefined) => void;
+  onChange: (value: string | number | undefined | string[]) => void;
 } & Omit<TextFieldProps, 'onChange'>) {
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
   if (parameter.type === 'source') return null;
 
   if (parameter.key === 'datasetId') {
@@ -30,49 +26,7 @@ export default function AgentInputField({
   }
 
   if (parameter.type === 'image') {
-    const handleFiles = async (file: File) => {
-      try {
-        const formData = new FormData();
-        formData.append('image', file);
-
-        const response = await uploadImage({ input: formData });
-        props.onChange(response.uploads.url);
-      } catch (error) {
-        Toast.error(error.message);
-      }
-    };
-
-    return (
-      <StringField
-        {...({ parameter } as any)}
-        size="small"
-        {...props}
-        multiline={false}
-        InputProps={{
-          endAdornment: (
-            <InputAdornment position="end" sx={{ mr: -0.75 }}>
-              <Stack direction="row" alignItems="center" gap={1}>
-                <input
-                  type="file"
-                  accept="image/*"
-                  style={{ display: 'none' }}
-                  ref={fileInputRef}
-                  onChange={async (e) => {
-                    const file = e.target.files?.[0];
-                    if (file) {
-                      handleFiles(file);
-                    }
-                  }}
-                />
-                <IconButton onClick={() => fileInputRef.current?.click()}>
-                  <AttachFileIcon sx={{ fontSize: 18 }} />
-                </IconButton>
-              </Stack>
-            </InputAdornment>
-          ),
-        }}
-      />
-    );
+    return <ImageField parameter={parameter} {...props} />;
   }
 
   const FIELDS: { [type in NonNullable<Parameter['type']>]?: ComponentType<any> } = {