+++ packages/ai-runtime/src/front/contexts/Debug.tsx
@@ -7,6 +7,10 @@ import { RuntimeOutputVariable } from '../../types';
 export interface DebugContextValue {
   agentId?: string;
   outputId?: string;
+  hoverOutputId?: string;
+  tabId?: string;
+  setTabId?: (id: string) => void;
+  setHoverOutputId: (id: string) => void;
   open: (options: { agentId?: string; output?: { id: string } | { name: RuntimeOutputVariable } }) => void;
   close: () => void;
 }
@@ -16,16 +20,19 @@ const debugContext = createContext<UseBoundStore<StoreApi<DebugContextValue>> | 
 export function DebugProvider({
   children,
   openSettings,
+  agentId: defaultAgentId,
 }: {
   children?: React.ReactNode;
   openSettings: (options: { agentId: string; output: { id: string } | { name: RuntimeOutputVariable } }) => {
     agentId?: string;
     outputId?: string;
   };
+  agentId?: string;
 }) {
   const [state] = useState(() =>
     create(
       immer<DebugContextValue>((set, get) => ({
+        agentId: defaultAgentId,
         open: ({ output, ...options }) => {
           const agentId = options.agentId || get().agentId;
           if (!agentId || !output) return;
@@ -38,8 +45,21 @@ export function DebugProvider({
         },
         close: () => {
           set((state) => {
-            state.agentId = undefined;
+            state.agentId = defaultAgentId;
             state.outputId = undefined;
+            state.hoverOutputId = undefined;
+          });
+        },
+        // hover Tab Show Output Component
+        setHoverOutputId: (id: string) => {
+          set((state) => {
+            state.hoverOutputId = id;
+          });
+        },
+        // hover Output Component Show Tab
+        setTabId: (id: string) => {
+          set((state) => {
+            state.tabId = id;
           });
         },
       }))
@@ -51,6 +71,36 @@ export function DebugProvider({
 
 export function useDebug<U>(selector: (state: DebugContextValue) => U): U | undefined {
   const ctx = useContext(debugContext);
+  if (!ctx) return undefined;
+
+  return ctx(selector);
+}
+
+export interface DebugDialogContextValue {
+  open?: boolean;
+  setOpen?: (open: boolean) => void;
+}
+
+const debugDialogContext = createContext<UseBoundStore<StoreApi<DebugDialogContextValue>> | null>(null);
+
+export function DebugDialogProvider({ children }: { children?: React.ReactNode }) {
+  const [state] = useState(() =>
+    create(
+      immer<DebugDialogContextValue>((set) => ({
+        setOpen: (open) => {
+          set((state) => {
+            state.open = open;
+          });
+        },
+      }))
+    )
+  );
+
+  return <debugDialogContext.Provider value={state}>{children}</debugDialogContext.Provider>;
+}
+
+export function useDebugDialog<U>(selector: (state: DebugDialogContextValue) => U): U | undefined {
+  const ctx = useContext(debugDialogContext);
   if (!ctx) return undefined;
 
   return ctx(selector);